Multiple updates:
- Use webpack for doing most packaging tasks instead of using a manual TS script for everything.
- Update webpack to automatically extract out the debugger package to ./daffodil-debugger-${daffodilVersion}-${extVersion}.
  - This allowed for the code to unzip the debugger in the extension to be removed.
  - This also allows for easier running of the debugger.
  - No longer extract the package out to somewhere else on the user system.
    - The extracted debugger folder with the bin and lib dirs are bundled with the extension source.
- Update scalawatch to build scala code then generate the version.ts, this will ensure the updated debugger package gets unzipped.
- Move location of webpack files as now there are two.
  - webpack/ext-dev.webpack.config.js - used when running locally.
  - webpack/ext-package.webpack.config.js - used when creating a package.
- Remove build/scripts directory.
- Move build/scripts/yarn-scripts.ts to build/yarn-scripts.ts
- Remove .vscodeignore, gets automatically created during the packaging to ignore node_modules.
- Remove Backend class and archiveUrl method from Artifact class
- Remove execSync commands from yarn-scripts.ts

Closes #692
Closes #711

+++ build/yarn-scripts.ts
@@ -17,19 +17,24 @@
 
 // @ts-nocheck <-- This is needed as this file is basically a JavaScript script
 //                 but with some TypeScript niceness baked in
+const path = require('path')
 const fs = require('fs')
 const glob = require('glob')
 const nodemon = require('nodemon')
 const concurrently = require('concurrently')
 
 function rmFileOrDirectory(path) {
-  if (fs.existsSync(path))
-    fs.rmSync(path, { recursive: true })
+  if (fs.existsSync(path)) fs.rmSync(path, { recursive: true })
 }
 
 function genVersionTS() {
-  const version = JSON.stringify(require('../../package.json').version).replace(/"/g, "'")
-  fs.writeFileSync('./src/version.ts', `/*
+  const version = JSON.stringify(require('../package.json').version).replace(
+    /"/g,
+    "'"
+  )
+  fs.writeFileSync(
+    './src/version.ts',
+    `/*
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.
@@ -47,13 +52,23 @@ function genVersionTS() {
  */
 
 export const LIB_VERSION = ${version}
-`)
+`
+  )
 }
 
-const nodeclean = () =>
-  ['out', 'dist'].forEach((dir) => rmFileOrDirectory(dir))
+const nodeclean = () => {
+  ;['out', 'dist'].forEach((dir) => rmFileOrDirectory(dir))
+  glob('daffodil-debugger-*', { cwd: '.' }, (err, files) => {
+    if (err) {
+      console.log(err)
+      return
+    }
 
-function scalaclean() {
+    files.forEach((dir) => rmFileOrDirectory(dir))
+  })
+}
+
+const scalaclean = () => {
   glob('**/target', { cwd: '.' }, (err, files) => {
     if (err) {
       console.log(err)
@@ -67,13 +82,41 @@ function scalaclean() {
 function watch() {
   concurrently(
     [
-      "yarn scalawatch",
-      "webpack --watch --devtool nosources-source-map --config ./build/extension.webpack.config.js",
-      "yarn watch:svelte"
+      'yarn scalawatch',
+      'webpack --watch --devtool nosources-source-map --config ./webpack/ext-dev.webpack.config.js',
+      'yarn watch:svelte',
     ],
     {
       killOthers: ['failure', 'success'],
     }
+  )
+}
+
+function package() {
+  const pkg_dir = 'dist/package'
+
+  // create .vscodeignore to not package all node_modules into the vsix
+  fs.writeFileSync(
+    path.join(pkg_dir, '.vscodeignore'),
+    `# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+**/node_modules/**/*
+!node_modules/@omega-edit/server/bin
+!node_modules/@omega-edit/server/lib
+`
   )
 }
 
@@ -82,4 +125,5 @@ module.exports = {
   nodeclean: nodeclean,
   scalaclean: scalaclean,
   watch: watch,
+  package: package,
 }
